:sparkles: Avoid calling the document store for image sets.
Using the document store to discover image model UUIDs for republishing is only applicable if the image set has been published successfully to at least one environment. Although simply generating the alternative UUID when it's not found directly may seem less robust, the probability that we find and republish the wrong piece of content in this case is vanishingly small.